Fixes for remote installations
As far as I am aware, this is insufficient to fix nwmTools, but I'm merging it now so that the rest of the progress is present.

HydroGOF: This was unceremoniously removed from CRAN some time ago, so it needs to be pulled from GitHub. HydroTSM is a dependency in the same boat.